+++ app/__init__.py
@@ -84,6 +84,7 @@ def create_app(config_object="config.Config"):
             "HAZARD_TYPES": models.HAZARD_TYPES,
             "HAZARD_SEVERITIES": models.HAZARD_SEVERITIES,
             "MAP_FEATURE_CATEGORIES": models.MAP_FEATURE_CATEGORIES,
+            "FIRE_RANKS": models.FIRE_RANKS,
         }
 
     return app